move hooks to componentHelpers module
